arg-cache: Do not leak an interface info structures on Callbacks
The cache is leaking the InterfaceInfo for each callback we pass as
BaseInfo is implicitly adding a new reference, but in the CallbackIn
case we were getting the interface from the type info, ignoring its
transfer-full return type.

To avoid this, just re-use the interface info we already got at
initialization time, removing the uneeded function call and fixing
the extra reference.