boxed: Templatize template classes to be re-usable for multiple types
We have multiple boxed types around but we have limited ability to
generate code out of them depending on the type they hold.

Instead, add CRTP to the whole boxed classes so that we can use generated
code that changes based on the type we need.

Rebased and edited to use CRTP by Philip.