boxed: Throw when trying to set a nested field value to a wrong type
In case a nested value is a boxed type we know how to handle, we should
to get the object as is and ensure that we're associating the same
types, otherwise we may end up generating a temporary object that will
have de-facto the expected interface info (as per how we initialize its
prototype).

Causing us being allowed to associate a field to whatever type.

Also avoid creating the prototype if the above conditions are and reduce
duplicated tests.