function: Use simpler GjsParamType pure array instead of vector
While vectors are cool, they implies some memory overhead that we may
not need in some cases.

In the GjsCallbackTrampoline in fact we already know the size of the
callable arguments, so we can just use a smart pointer of GjsParamType[]
and use native features.

(Changed to std::unique_ptr<T[]> by Philip)