build: disable sysprof agent for subproject fallback
Currently disabled by default in Sysprof (but to be re-enabled soon) is
a new -Dagent=true build option. This is used to allow profiling within
containers and installed in the container. However, it requires compiling
more of Sysprof such as libsysprof and that is unecessary for various CI
usage of Sysprof.

This disables it as we don't need it in GJS.