maint: Fix IWYU includes in files that previously failed
Due to the bug fixed by the previous commit, many files' includes had
gotten out of sync. This fixes all of the IWYU complaints.

We need to add some lines to the mapping file, either due to an IWYU
upgrade or changes in the header files; I'm not sure. It's annoying that
we have to list all the glib/*.h files separately, but that is for two
reasons: there is no glob for glib/* that can leave out <glib/gstdio.h>
and other public GLib headers; and glib/galloca.h should redirect to the
system alloca.h.