CI: Temporarily disable --werror
There are some warnings showing up with the new version of the compiler
in the newly-built image that don't have obvious fixes. Normally we stop
the build if there are warnings, but we will temporarily disable that and
allow it to proceed until these new warnings can be fixed. (One of them
seems to come from SpiderMonkey and would be fixed in the SpiderMonkey
upgrade, it seems.)

(cherry picked from commit 825d85a1057d2330c2811116315669606b8e4ac3)