console: Fix setjmp warnings
These variables now generate warnings that they "might be clobbered by
'longjmp' or 'vfork'". Declaring them volatile prevents this.

As per https://stackoverflow.com/a/2105840/172999, the C99 standard says
"the values of objects of automatic storage duration that are local to the
function containing the invocation of the corresponding setjmp macro that
do not have volatile-qualified type and have been changed between the
setjmp invocation and longjmp call are indeterminate."