boxed: Never override methods with fields
In case an union or struct field has the same name of a method that
clashes with a field, we should give priority to the method name, and
use a '_' prefixed name for object properties that represent fields.

This fixes the case of ClutterEvents that have a `type()` method and a
`type` field, that is now renamed to `_type` by default.