Gio: Promisify GDBusProxy.init_async by default
Some code may promisify it after that it has been overridden, this means
that when called we may not use the original method correctly as the
function won't have the arguments set anymore and thus its length will
be 0, causing the assumption included in commit 58840261 not to be true
anymore.

Since the function is already a wrapper, we can safely override it to be
a promise and then override it again. Allowing code simplifications and
being more future-proof, when all the async functions will be promises
by default.

Fixes: #494