function: Add missing include
This one is tricky, and not caught by IWYU, because constructing
JS::RootedValueVector requires computing sizeof(JS::Value). It was also
not caught in CI because the build error doesn't occur when using
precompiled headers, and all of the CI build jobs currently do.

Work around it by writing out the template parameter explicitly, instead
of relying on the typedef.

Closes: #503