function: Do not allow returning a transfer-none string from JS callback
In order to marshal a string from JS into C, we have to create a temporary
UTF-8 string. That string cannot be returned as transfer-none, because
there is no obvious place to keep it alive so that it can live as long as
the C code needs it to. See #519 for more explanation.

This is currently "supported" but causes a memory leak. Instead, throw
when creating a callback or vfunc in JS that would otherwise have this
memory leak.