context: Refactor warning about unhandled promise rejections
Move the exit-time checks for unhandled promise rejections into a single
place in context.cpp, a GjsContextPrivate::exit_immediately() method which
takes care of that (and in future, any other actions that have to be
performed on a dirty exit.) These checks were scattered in several places.

Now we warn about unhandled promise rejections when exiting, and at the
end of draining the microtask queue.