function: Always initialize callback return value
When callback_closure() exits early, for example due to being called
during GC, the return value would not be initialized. This value is
often non-zero. If the callback is a source func of an idle or a timeout
this would then get interpreted as G_SOURCE_CONTINUE and the same would
repeat in the next iteration. If this happens fast enough, this results
in the entire process being seemingly frozen while spamming the log with
error messages.

To fix this always initialize the return value to 0 or a comparable
neutral value.

Related: https://gitlab.gnome.org/GNOME/gnome-shell/-/issues/1868