gi/arg: Handle GValue's flat-arrays just like any other flat-array
GValue has nothing special a part being more widely used, so we can just
use the same code path to manage GVariant arrays that are not very common
anyways.

The previous codepath did not include particular optimizations, so we
can just rely on generic code, so we have less to maintain and can handle
more structures.