gi/boxed: Do not leak all the GVariant instances
Due to the misleading transfer annotation in GIMarshallingTests we were
leaking all the GVariant values that were handled by gjs because for each
one we were adding 2 references, and removing just one.

The problem was that the GIMarshallingTests.array_gvariant_full_in()
function was actually returning a floating GVariant value (and so to be
considered as transfer none) and so its transfer was not full (but rather
container, if any).

Given that it's not possible to know at call time whether a function returns
a floating reference or not, those should be considered as transfer-none by
default (and so GLib does).

This reverts commit 9fee568d832acd4907225aeca40ed0c97ebdd96c.

Closes: https://gitlab.gnome.org/GNOME/gjs/-/issues/499
See: https://gitlab.gnome.org/GNOME/gobject-introspection/-/merge_requests/399