context: Add gjs_context_run_in_realm() API
For specialized cases where the caller needs to mess around with
SpiderMonkey API directly, we offer the gjs_context_get_native_context()
API which you can cast to the JSContext*. However, nowadays in
SpiderMonkey you cannot do much without having entered a JS::Realm,
which we did not expose.

In order to avoid an API where we expose another void pointer and rely
on the caller to cast it, we add gjs_context_run_in_realm(), which
enters the GJS main realm, calls the given callback in which you can do
things with JSAPI, and then leaves the realm.

Closes: #536