jsapi-util: Remove outdated comment
We actually don't need to use JS::GetExceptionCause here, because
setting an error's `cause` property doesn't update the value in the
Error object's reserved slot. So, in the situation where it's possible
to cause a side effect by overwriting the property with an accessor, we
can already no longer avoid triggering the side effect by reading the
reserved slot.