gi/args.cpp: Fix build with Visual Studio
Apparently Visual Studio is more strict about the using the results of a
bitwise-or as a parameter to a function that takes an enumeration.

So, make the overloaded operator| in arg-cache.cpp shared, by defining a
prototype in arg.h, and moving up the implementation in arg-cache.cpp up
a bit so that things don't become ambiguous in arg-cache.cpp.