cwrapper: Define constructor object on current global
If not passing in a module object on which to define the constructor
object, CWrapper::create_prototype() uses the global object. It doesn't
make any difference currently, but it should use the current global
instead of the main realm's global in the case where it was called from a
different realm.