context: Remove gjs_get_import_global()
It has a confusing name, and there is another method,
GjsContextPrivate::global(), that does exactly the same thing.

Some renames in coverage.cpp and debugger.cpp now that we don't have to
use a temporary variable "debuggee" to hold the result of the long
expression gjs_get_import_global(cx); actually "debuggee" and
"debuggee_wrapper" were the same object.