arg-cache: Properly handle inout array length arguments
The inout length arguments were not properly handled by our code because
we were trying to access to the length value as if the GIArgument value
was set to it, but we should had actually read the pointed value.

We also now asserting if the length parameter direction does not match
the array one, this may technically not happen all the times, but we
always assumed it in code, and at least now we can fail early if this
happens.

As per this enabling tests that were previously failing.

Closes: https://gitlab.gnome.org/GNOME/gjs/-/issues/95