arg-cache: Add Implementations for zero-terminated and fixed-length arrays
These arrays may requires to be free their content and this does not
happen when using the GenericIn and GenericInOut handlers because they
may end up clearing only the array and not its content.

So use specific argument cache implementations in these cases, handling
the cleanup properly.

This is already tested but we did not test the case for function
accepting nullable zero-terminated arrays, so add some cases for them.

Closes: https://gitlab.gnome.org/GNOME/gjs/-/issues/561