value: Load signal details only once for a signal invocation
We were trying to get the signal information for each argument multiple
times instead of just passing all the known information to the
conversions functions to avoid repeating the same if needed.

So, just keep around the argument infos that we were already fetching
during the values conversion and pass them together with the signal info
to the conversion functions so that they can take advantage of them if
they need to without having to repeat the same steps.