value: Handle gvalues holding a null-pointer as first thing
GValues with pointers may contain null values but we don't guard against
them in all the cases, so avoid specific logic and just check once for
all if a value is null and just return the null object in case.

GStrv is the only special case we have in which an empty array should be
returned instead.

Also here we could do better checking using the signal information we
may get, but sadly that's not something we can manage without breaking
most cases, because signals are often coming with no such info.