gjs: Add simple class to wrap a C++ object inside a JS Object
We may often need to use JS APIs to pass C++ data around and it requires
using JSValues for that.

So add a simple class that allows to easily create a JSObject that can
be used as private data and that contains c++ data that can be easily
destroyed when gc happens.

Added tests to check that things work as expected.