object: Bind property specs to getters/setters removing property cache
In order to get and set properties we used a cache to associate each
property to its pspec, as we were only referring to the property ID when
we were handling the property in the getter implementations.

However we can avoid this by instead binding to each property
getter/setter pair the related GParamSpec instead.

We do this using a simple wrapper instance without destroy-notify, saving
an extra pointer per property. GParamSpecs are owned by GObjectClass, that
is basically never finalized. In particular it can never happen while an
object is alive (and so when we could make use of its
param specs).

The only case in which it's technically possible that param specs are
deleted are before an object is initialized, so refactoring this phase so
that if we are creating an object with properties on construction, the
class is kept around until we've not created it.