object: Bind GIFieldInfo to each field property
For every object we were using a fields cache, in order to speedup the
access to these values, providing them the GI field info data.

However, this can now achieved in a simpler way and with less memory
usage by using a simple wrapper bound to the object property that we
define.

In this way we set the field info only once and we keep it alive with
the object (as we did before), but the access to it is faster because
we don't have to go through any map.