object: Add support for using property setters and getters
Objects may define object setters and getters and they provide some
benefits because we can avoid going through GValue and just call the
proper function to get or set the same value. Potentially also avoiding
copies.

To do this, implement a different kind of property setters and getters
depending on what the type supports.

Use GI to call property getters and setters, as this is more optimized,
even though we have to manually handle a few more things.