arg-cache/function: Simplify the return parsing for basic types
When handling basic return types functions can avoid inspecting the
typeinfo and even remembering it in the argument cache, as we can just
do everything from the type tag.

So, simplify Arg::NumericReturn to only remember the tag is constructed
with and at call function time we now use the GTypeInfo only if
required.

This both saves memory because for simple types we store only a minimal
amount in cache, and reduces the computation because once we've a type
definition we go through it with less checks.