object: Reuse the same property descriptors (and associated objects)
An object properties may have various aliases, in particular we support
properties to have dashed-names, camelCaseNames and param_names.

Now, till now for each of these properties we were creating new property
descriptors (and in particular function objects with associated data)
and so this come to a memory cost.

To avoid this we now create just a single property descriptor that is
always named using the property canonical name and we re-use it for all
the aliases.

So:
 - If the property name we're handling is not using the canonical name
   (i.e. the dashed-name), we now look for a canonical property
   description
 - If one is found, then we just reuse it for the new property.
 - If none is found, we continue defining the property but using both
   the property ids, still sharing the same descriptor.

While the property descriptor allocation can't probably be shared
internally, the setter and getter objects (and related data) can.