value: Use direct access to GValue field elements
GValue elements can often be accessed directly, avoiding function calls
overhead.

It's something that GLib does in various implementations, and also does
it with generated code.

Now, gjs can also do this in its internal code since accessing the
values directly can be considered safe in the internal code and this
has a good effect on performances.

The safety of this is everywhere ensured by the fact that we check on
the value GType before accessing it or we explicitly set it.