arg-cache: Scalar out parameters should not be caller-allocates
The (out caller-allocates) annotation is for structured types that need to
be allocated for the function to fill in. Caller-allocates for scalar
types is a mistake, so mark the function non-introspectable because
otherwise it will return the value of the pointer.

If needed, in the future we could add another argument cache marshaller to
handle this, but let's see if we can get all the faulty annotations fixed
upstream first.

Closes: #570