gi: Allow boxed closures to be passed to GI
Normally, when passing a closure to a GI function or returning one from
a vfunc, you want to return a JavaScript function. However, there are
cases where JS code may need to deal with boxed GClosures instead.

Previously, if you tried to pass a boxed closure, GJS would assume it's
a callable object and wrap it in a GJS closure. Then, when that closure
is called, it would try to invoke the boxed wrapper as a JS function
instead of invoking the closure it wraps.

Now, the translation code detects when you pass a boxed closure to a
closure argument, and passes the boxed closure directly through.