gi/private: Add associateClosure() function
This function takes a GObject and a JavaScript function, creates a
GClosure, and associates that GClosure with the object using
ObjectInstance->associate_closure(). It returns the GClosure as a boxed
instance.

Normally, when you pass a function to a GI method, it is wrapped in a
GJS-provided implementation of GClosure. This closure keeps the
JavaScript function alive until the closure is invalidated. If the
closure is never invalidated, the function stays alive indefinitely.

This behavior is correct. Imagine connecting a signal to the function
`() => console.log("Hello, world!")`. One would expect the signal to
print "Hello, world!" whenever it is emitted until it is disconnected
and the closure disposed.

Now imagine connecting a signal to another function:

    let obj = new MyGObject();
    otherObj.connect("some-event", () => obj.printHelloWorld());

One would still expect the signal handler to remain until it is
disconnected or otherObj is destroyed. The code does nothing to indicate
that it should be disconnected, so the function must be kept alive, and
because it refers to `obj`, that must be kept alive as well.

However, while this behavior is *correct,* it is not always *intended.*
There are many cases where we want the signal to remain connected only
for the lifetime of a given object. This was not previously possible for
two reasons:

1. You can't override vfunc_dispose() and disconnect signal handlers
there, because dispose might run during GC.

2. It can cause a reference cycle. Imagine connecting a signal like
this:

    let obj = new MyGObject();
    obj.connect("some-event", () => obj.printHelloWorld());

Like before, this code does nothing to indicate that the signal handler
should be disconnected. Because it is not disconnected, the function
stays alive, and because the function refers to obj, obj stays alive,
keeping the signal connected indefinitely. The GC cannot detect this
loop because the JS function is rooted until the closure is disposed,
which will not happen because it is referred to by the object, which is
referred to by the rooted function.

What is needed, then, is a way to explicitly tell GJS which object to
tie the closure's lifetime to. This is what associateClosure() does. It
uses some existing code that traces the JS function through an object
wrapper. That way, it doesn't have to be rooted to stay alive, and can
be garbage collected with the object. Now instead of the closure keeping
the function alive, the object invalidates the closure (through C code)
when it is disposed.