modules: gtk: Limit lifetime of signal handlers
When connecting template signals, limit the lifetime of the handler
function to that of the current object, using the mechanisms added in
the previous commits.

This matches the behavior of GtkCBuilderScope, which uses
g_cclosure_new_object() which disconnects the signal handler when the
template object is disposed. Without this behavior, as explained in
previous commit messages, the handler and template object are leaked via
reference cycle.