heapgraph: Read heap in two passes
Since many of the strings are now bunched up at the end of the file,
we'll need to read the heap file in two passes, one to read all the
nodes and one to read all the edges. The previous assumption that an
edge line would occur after the target no longer holds.