overrides: Make class object a parameter of register type hooks
The GObject register type hooks previously used their this-object as the
class object to act upon. That was awkward because in many cases they
weren't called in such a way that the correct class object was already
the this-object, making it necessary to use .call(). Instead, we can
make the class object the parameter.