Gtk3: Fix leak in GtkBuilder template signal connections
The widget class's GtkBuilderConnectFunc would previously capture the
newly created object instance, making it unable to be collected.

It's not actually necessary for the GtkBuilderConnectFunc to capture an
external reference to the object instance. It can be obtained from
Gtk.Builder.prototype.get_objects() because for a template, there will
always be one of the created objects that is of the template type.

There is still a leak in Gtk4 template signal connections, see !891. The
test cases are adapted from there, and are written by James Westman.