Gtk3: Set GtkBuilderConnectFunc in class init
We were setting and re-setting the GtkBuilderConnectFunc each time an
instance was constructed, instead of at class init time as the
documentation says: "Note that gtk_widget_class_set_connect_func() must
be called from a composite widget class's class initializer after
calling gtk_widget_class_set_template()."

Especially now that we are not keeping a reference to the constructed
instance, it's easy to move this into the class init function. However,
we need to do it in two places, since the legacy Lang.Class also relied
on it happening in the instance init function. So, it becomes easiest to
move it into _common.js.