jsapi-util: Refactor gjs_format_stack_trace()
A better name for this function is format_saved_frame(), since it formats
a JSAPI SavedFrame object. Refactor the conversion to filename encoding
out of this function, so that we only re-encode when we are about to print
out the string. (We also use it for tracking unhandled promise rejection
call sites, so there is no need to do the encoding step when we are using
it for that purpose.)