object: Fix introspection info types
Some of the introspection info types were not correct, but this was not
a problem because all of the info types are typedefs for GIBaseInfo.
However, in gobject-introspection-2.0 this will become a problem.

resolve_on_interface_prototype() needs a GIInterfaceInfo.
gjs_lookup_object_prototype_from_info() and
gjs_lookup_object_constructor_from_info() can take either a GIObjectInfo
or GIInterfaceInfo, so give them a GIBaseInfo pointer and do a type
check.