arg: Separate gjs_arg_member template for long/ulong
On our usual platform, int64_t is a typedef for long, so
gjs_arg_set<long> would set the v_int64 member of GIArgument instead of
v_long. Technically not correct but probably always works.

On macOS, int64_t is a typedef for long long, so gjs_arg_set<long> would
actually not compile.

This adds an extra template where we can define our own tags that are
not part of the GITypeTag enum. (There is no GI_TYPE_TAG_LONG that would
be suitable for this template specialization.)