gi: Add null-safe C++ wrappers for libgirepository
The APIs from girepository are used everywhere, are verbose, and are not
always clear about what return values are owned and when they can be null.

Some of these things are fixed in libgirepository 2.0. Since we are going
to switch to that anyway, let's add a wrapper (in gi/info.h) that encodes
these things directly in the API.

The null-safe API is explained in comments at the top of gi/info.h.
AutoFooInfo is owned, FooInfo is unowned (and should be passed around to
functions as const FooInfo) and there are also a limited number of stack-
only introspection info classes. There is a certain amount of C++ blah
blah blah needed (made worse by the fact that in libgirepository 1.0, all
the structs are just typedefs of GIBaseInfo) but it largely remains
confined to gi/info.h.