foreign: Remove outdated comment
We don't have access to the GITypeInfo here at all. We do have the
GIStructInfo when looking up the foreign marshalling info, but we don't
have it at registration time when the marshalling info is added to the
hash table. So there's no reason to have this as a FIXME.