foreign: C++ify the data structures
Cpplint doesn't like my braced initializer when string-concatenating a
unique key for the struct info? Well how about we dispense with the
string-concatenation and instead use a map with a key type of
std::pair<namespace, name> and compute a hash code.

Hopefully, this makes the code more typesafe and easier to read.