meson: fix automagic dependency lookup for cairo
Using custom required: conditions based on msvc fallbacks means that
cairo can be found whenever it is installed, even when disabled. This
breaks logic checks later on, and produces wasteful lookups when force
disabled anyways.

Fix this by using the same logic meson would use with a true feature
option, and cause cairo to always be not-found when it is disabled,
using a guaranteed blank / not-found dependency.