CI: Build newer GLib in debug Docker image
We are currently falling between the cracks into a situation where the
responsibility of building the GLib GIRs is being transferred from
gobject-introspection to GLib. On the Docker image we have a g-i that is
new enough to require a newer GLib, which the build pulls in via a
subproject. However, building GIRs in GLib-as-a-subproject is not yet
supported, so the GIRs don't get built properly.

We solve this by building a new enough GLib and installing it on the
system, so it won't be pulled in as a subproject.

The non-debug image doesn't need to be updated because there we just use
the system's gobject-introspection.