GObject: Throw on refcounting methods
While we're at it, may as well throw on refcounting methods too. We've
talked about it for a long time, so maybe we can land this early in the
GNOME 47 cycle and revert it if it causes problems.

The refcounting methods are used in the GObject refcounting tests, so we
need to replace them there with special functions on GjsTestTools.

Also inspired by PyGObject:
https://gitlab.gnome.org/GNOME/pygobject/-/blob/e51fdc963e42a603e73036a24b2b94b7b403f778/gi/overrides/GObject.py#L531-534