debugger: Refactor throwOrReturn to deal with exceptions from eval
Usually if we are evaluating an expression inline for it to be used in a
command, we don't want an exception occurring from the evaluation to
interfere with the control flow of the program. We just want to print and
swallow the exception.

This adds a new function evalInFrame() which evaluates the given
expression inside the current frame, with debugger bindings. Refactor
throwOrReturn() to use this function.

Aside from dealing nicely with exceptions, this adds the feature of being
able to use debugger bindings with the return and throw commands, like
'return $3'. Add tests for these new features.